Add Xcode-only functional tests for different QoS levels (requires server)
